added components for cypto and crypto container
